feat: add skeleton loader to query layers while waiting for remote configs
