+++ src/RemoteConfig.jsx
@@ -1,5 +1,6 @@
-import { activate, fetchConfig, getRemoteConfig } from 'firebase/remote-config';
+import { fetchAndActivate, getRemoteConfig } from 'firebase/remote-config';
 import PropTypes from 'prop-types';
+import { BulletList } from 'react-content-loader';
 import { RemoteConfigProvider, useInitRemoteConfig } from 'reactfire';
 import defaultConfig from './remote_config_defaults.json';
 
@@ -13,11 +14,9 @@ export default function RemoteConfig({ children }) {
         fetchTimeoutMillis: 10000,
       };
 
-      // immediately activate the latest cached config to help with initial load performance
-      await activate(remoteConfig);
-
-      // fetch new configs for the next load
-      fetchConfig(remoteConfig);
+      await fetchAndActivate(remoteConfig);
+      // use for testing loader...
+      // await new Promise((resolve) => setTimeout(resolve, 3000));
 
       return remoteConfig;
     }
@@ -26,7 +25,7 @@ export default function RemoteConfig({ children }) {
   if (status === 'loading') {
     console.log('loading remote config');
 
-    return <span>getting configs...</span>;
+    return <BulletList className="p-2" />;
   }
 
   console.log('remote config loaded');
