+++ src/functions/update.test.js
@@ -1,6 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { fieldNames } from '../config';
-import { arraysToObjects, getLinksConfig } from './update';
+import { arraysToObjects } from './update';
 
 describe('arraysToObjects', () => {
   it('converts an array of arrays to an array of objects', () => {
@@ -27,22 +26,5 @@ describe('arraysToObjects', () => {
       { name: 'Mary', age: 25 },
     ];
     expect(arraysToObjects(arrays, ['skip'])).toEqual(objects);
-  });
-});
-
-describe('getLinksConfig', () => {
-  it('gets the links config', () => {
-    const values = [
-      [fieldNames.links.id, fieldNames.links.description, fieldNames.links.url],
-      ['1', 'description 1', 'url 1'],
-    ];
-    const expected = {
-      1: {
-        description: 'description 1',
-        url: 'url 1',
-      },
-    };
-
-    expect(getLinksConfig(values)).toEqual(expected);
   });
 });
